Update src/Core/Context.cpp
Remove this redundant cast.

+++ src/Core/Context.cpp
@@ -125,7 +125,7 @@ void Context::Update() {
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
     constexpr ms_t frameTime =
-        FPS_CAP != 0 ? static_cast<float>(1000.0F / FPS_CAP) : 0;
+        FPS_CAP != 0 ? 1000.0F / FPS_CAP : 0;
     ms_t afterUpdate = Util::Time::GetElapsedTimeMs();
     ms_t updateTime = afterUpdate - m_BeforeUpdateTime;
     if (updateTime < frameTime) {
